+++ src/service/ValClient/Contract.js
@@ -1,5 +1,5 @@
 //import
-const AxiosClient = require('../resources/AxiosClient');
+const AxiosClient = require('../../resources/AxiosClient');
 
 //service
 class Contract {
@@ -17,7 +17,7 @@ class Contract {
     }
 
     /**
-    * @param {string} puuid PlayerUUID
+    * @param {String} puuid PlayerUUID
     */
     async Fetch(puuid) {
         const response = await this.AxiosClient.get(this.Region.url.playerData + `/contracts/v1/contracts/${puuid}`);
@@ -26,8 +26,8 @@ class Contract {
     }
 
     /**
-    * @param {string} puuid PlayerUUID
-    * @param {string} contractId ContractID
+    * @param {String} puuid PlayerUUID
+    * @param {String} contractId ContractID
     */
     async Activate(puuid, contractId) {
         const response = await this.AxiosClient.post(this.Region.url.playerData + `/contracts/v1/contracts/${puuid}/special/${contractId}`);