+++ src/service/ValClient/Match.js
@@ -1,5 +1,5 @@
 //import
-const AxiosClient = require('../resources/AxiosClient');
+const AxiosClient = require('../../resources/AxiosClient');
 
 //service
 class Match {
@@ -10,7 +10,7 @@ class Match {
 
     /**
     * @description Get contract definitions
-    * @param {string} matchId MatchID
+    * @param {String} matchId MatchID
     */
      async FetchMatchDetails(matchId) {
         const response = await this.AxiosClient.get(this.Region.url.playerData + `/match-details/v1/matches/${matchId}`);
@@ -19,10 +19,10 @@ class Match {
     }
 
     /**
-    * @param {string} puuid PlayerUUID
-    * @param {string} queueId QueueID
-    * @param {number} startIndex startIndex
-    * @param {number} endIndex endIndex
+    * @param {String} puuid PlayerUUID
+    * @param {String} queueId QueueID
+    * @param {Number} startIndex startIndex
+    * @param {Number} endIndex endIndex
     */
      async FetchMatchHistory(puuid, queueId = null, startIndex = 0, endIndex = 10) {
         let _url = this.Region.url.playerData + `/match-history/v1/history/${puuid}?startIndex=${startIndex}&endIndex=${endIndex}` ;