+++ src/service/ValClient/Pregame.js
@@ -1,5 +1,5 @@
 //import
-const AxiosClient = require('../resources/AxiosClient');
+const AxiosClient = require('../../resources/AxiosClient');
 
 //service
 class Pregame {
@@ -9,7 +9,7 @@ class Pregame {
     }
 
     /**
-    * @param {string} puuid PlayerUUID
+    * @param {String} puuid PlayerUUID
     */
      async GetPlayer(puuid) {
         const response = await this.AxiosClient.get(this.Region.url.partyService + `/pregame/v1/players/${puuid}`);
@@ -18,7 +18,7 @@ class Pregame {
     }
 
     /**
-    * @param {string} matchId MatchID
+    * @param {String} matchId MatchID
     */
      async GetMatch(matchId) {
         const response = await this.AxiosClient.get(this.Region.url.partyService + `/pregame/v1/matches/${matchId}`);
@@ -27,7 +27,7 @@ class Pregame {
     }
 
     /**
-    * @param {string} matchId MatchID
+    * @param {String} matchId MatchID
     */
      async GetMatchLoadouts(matchId) {
         const response = await this.AxiosClient.get(this.Region.url.partyService + `/pregame/v1/matches/${matchId}/loadouts`);
@@ -36,7 +36,7 @@ class Pregame {
     }
 
     /**
-    * @param {string} matchId MatchID
+    * @param {String} matchId MatchID
     */
      async QuitMatch(matchId) {
         const response = await this.AxiosClient.post(this.Region.url.partyService + `/pregame/v1/matches/${matchId}/quit`);
@@ -45,8 +45,8 @@ class Pregame {
     }
 
     /**
-    * @param {string} matchId MatchID
-    * @param {string} agentId CharacterID
+    * @param {String} matchId MatchID
+    * @param {String} agentId CharacterID
     */
      async SelectCharacter(matchId, agentId) {
         const response = await this.AxiosClient.post(this.Region.url.partyService + `/pregame/v1/matches/${matchId}/select/${agentId}`);
@@ -55,8 +55,8 @@ class Pregame {
     }
 
     /**
-    * @param {string} matchId MatchID
-    * @param {string} agentId CharacterID
+    * @param {String} matchId MatchID
+    * @param {String} agentId CharacterID
     */
      async LockCharacter(matchId, agentId) {
         const response = await this.AxiosClient.post(this.Region.url.partyService + `/pregame/v1/matches/${matchId}/lock/${agentId}`);