+++ src/service/ValClient/Store.js
@@ -1,5 +1,5 @@
 //import
-const AxiosClient = require('../resources/AxiosClient');
+const AxiosClient = require('../../resources/AxiosClient');
 
 //service
 class Store {
@@ -9,8 +9,8 @@ class Store {
     }
 
     /**
-    * @param {string} puuid PlayerUUID
-    * @param {string} itemTypeId ItemTypeID
+    * @param {String} puuid PlayerUUID
+    * @param {String} itemTypeId ItemTypeID
     */
      async GetEntitlements(puuid, itemTypeId) {
         const response = await this.AxiosClient.get(this.Region.url.playerData + `/store/v1/entitlements/${puuid}/${itemTypeId}`);
@@ -28,7 +28,7 @@ class Store {
     }
 
     /**
-    * @param {string} puuid PlayerUUID
+    * @param {String} puuid PlayerUUID
     */
      async GetStorefront(puuid) {
         const response = await this.AxiosClient.get(this.Region.url.playerData + `/store/v2/storefront/${puuid}`);
@@ -37,7 +37,7 @@ class Store {
     }
 
     /**
-    * @param {string} puuid PlayerUUID
+    * @param {String} puuid PlayerUUID
     */
      async GetWallet(puuid) {
         const response = await this.AxiosClient.get(this.Region.url.playerData + `/store/v1/wallet/${puuid}`);