+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.4.4 (March 23, 2022)
+
+- New Method For "Auth.Account" And "Auth.Multifactor"
+- API Client Can Return Error
+- Change "ValClient.Region" To "ValClient.region"
+- A little changes
+
 # 1.4.3 (March 22, 2022)
 
 - RiotApi Settings
@@ -5,8 +12,8 @@
 
 # 1.4.2 (March 21, 2022)
 
-- remove "AxiosData" from ValClient/Service class
-- more function in ValClient/Service
+- Remove "AxiosData" from ValClient/Service class
+- More function in ValClient/Service
 
 # 1.4.1 (March 21, 2022)